+++ test/td_dd_web/controllers/xlsx_controller_test.exs
@@ -38,7 +38,7 @@ defmodule TdDdWeb.DataStructures.XLSXControllerTest do
     end
 
     @tag authentication: [role: "admin"]
-    test "downloads xlsx content using scroll to search", %{conn: conn} do
+    test "downloads xlsx content for published notes using scroll to search", %{conn: conn} do
       df_content = %{"string" => %{"value" => "foo", "origin" => "user"}}
 
       domain = CacheHelpers.insert_domain()
@@ -92,7 +92,292 @@ defmodule TdDdWeb.DataStructures.XLSXControllerTest do
       end)
 
       assert %{resp_body: body} =
-               post(conn, Routes.xlsx_path(conn, :download, %{download_type: "editable"}))
+               post(
+                 conn,
+                 Routes.xlsx_path(conn, :download, %{
+                   download_type: "editable",
+                   note_type: "published"
+                 })
+               )
+
+      assert {:ok, workbook} = XlsxReader.open(body, source: :binary)
+      assert {:ok, [headers | content]} = XlsxReader.sheet(workbook, dsv.type)
+
+      assert headers == [
+               "external_id",
+               "name",
+               "tech_name",
+               "alias_name",
+               "link_to_structure",
+               "domain",
+               "type",
+               "path",
+               "string"
+             ]
+
+      assert content == [
+               [
+                 structure.external_id,
+                 dsv.name,
+                 dsv.name,
+                 "",
+                 "",
+                 domain.name,
+                 dsv.type,
+                 "",
+                 "foo"
+               ]
+             ]
+    end
+
+    @tag authentication: [role: "admin"]
+    test "downloads xlsx content for draft notes using scroll to search", %{conn: conn} do
+      df_content = %{"string" => %{"value" => "foo", "origin" => "user"}}
+
+      domain = CacheHelpers.insert_domain()
+
+      structure =
+        insert(:data_structure,
+          domain_ids: [domain.id],
+          draft_note:
+            build(:structure_note,
+              df_content: df_content,
+              status: :draft,
+              data_structure: nil
+            )
+        )
+
+      dsv =
+        insert(:data_structure_version,
+          data_structure: structure
+        )
+
+      %{id: id} =
+        CacheHelpers.insert_template(%{
+          name: dsv.type,
+          scope: "dd",
+          content: [
+            %{
+              "name" => "group",
+              "fields" => [
+                %{
+                  "name" => "string",
+                  "type" => "string",
+                  "label" => "Label foo"
+                }
+              ]
+            }
+          ]
+        })
+
+      insert(:data_structure_type, name: dsv.type, template_id: id)
+
+      assert :ok = StructureEnricher.refresh()
+
+      ElasticsearchMock
+      |> expect(:request, fn _, :post, "/structures/_search", _, opts ->
+        assert opts == [params: %{"scroll" => "1m"}]
+        SearchHelpers.scroll_response([dsv])
+      end)
+      |> expect(:request, fn _, :post, "/_search/scroll", body, [] ->
+        assert body == %{"scroll" => "1m", "scroll_id" => "some_scroll_id"}
+        SearchHelpers.scroll_response([])
+      end)
+
+      assert %{resp_body: body} =
+               post(
+                 conn,
+                 Routes.xlsx_path(conn, :download, %{
+                   download_type: "editable",
+                   note_type: "non_published"
+                 })
+               )
+
+      assert {:ok, workbook} = XlsxReader.open(body, source: :binary)
+      assert {:ok, [headers | content]} = XlsxReader.sheet(workbook, dsv.type)
+
+      assert headers == [
+               "external_id",
+               "name",
+               "tech_name",
+               "alias_name",
+               "link_to_structure",
+               "domain",
+               "type",
+               "path",
+               "string"
+             ]
+
+      assert content == [
+               [
+                 structure.external_id,
+                 dsv.name,
+                 dsv.name,
+                 "",
+                 "",
+                 domain.name,
+                 dsv.type,
+                 "",
+                 "foo"
+               ]
+             ]
+    end
+
+    @tag authentication: [role: "admin"]
+    test "downloads xlsx content for pending-approval notes using scroll to search", %{conn: conn} do
+      df_content = %{"string" => %{"value" => "foo", "origin" => "user"}}
+
+      domain = CacheHelpers.insert_domain()
+
+      structure =
+        insert(:data_structure,
+          domain_ids: [domain.id],
+          pending_approval_note:
+            build(:structure_note,
+              df_content: df_content,
+              status: :pending_approval,
+              data_structure: nil
+            )
+        )
+
+      dsv =
+        insert(:data_structure_version,
+          data_structure: structure
+        )
+
+      %{id: id} =
+        CacheHelpers.insert_template(%{
+          name: dsv.type,
+          scope: "dd",
+          content: [
+            %{
+              "name" => "group",
+              "fields" => [
+                %{
+                  "name" => "string",
+                  "type" => "string",
+                  "label" => "Label foo"
+                }
+              ]
+            }
+          ]
+        })
+
+      insert(:data_structure_type, name: dsv.type, template_id: id)
+
+      assert :ok = StructureEnricher.refresh()
+
+      ElasticsearchMock
+      |> expect(:request, fn _, :post, "/structures/_search", _, opts ->
+        assert opts == [params: %{"scroll" => "1m"}]
+        SearchHelpers.scroll_response([dsv])
+      end)
+      |> expect(:request, fn _, :post, "/_search/scroll", body, [] ->
+        assert body == %{"scroll" => "1m", "scroll_id" => "some_scroll_id"}
+        SearchHelpers.scroll_response([])
+      end)
+
+      assert %{resp_body: body} =
+               post(
+                 conn,
+                 Routes.xlsx_path(conn, :download, %{
+                   download_type: "editable",
+                   note_type: "non_published"
+                 })
+               )
+
+      assert {:ok, workbook} = XlsxReader.open(body, source: :binary)
+      assert {:ok, [headers | content]} = XlsxReader.sheet(workbook, dsv.type)
+
+      assert headers == [
+               "external_id",
+               "name",
+               "tech_name",
+               "alias_name",
+               "link_to_structure",
+               "domain",
+               "type",
+               "path",
+               "string"
+             ]
+
+      assert content == [
+               [
+                 structure.external_id,
+                 dsv.name,
+                 dsv.name,
+                 "",
+                 "",
+                 domain.name,
+                 dsv.type,
+                 "",
+                 "foo"
+               ]
+             ]
+    end
+
+    @tag authentication: [role: "admin"]
+    test "downloads xlsx content for rejected notes using scroll to search", %{conn: conn} do
+      df_content = %{"string" => %{"value" => "foo", "origin" => "user"}}
+
+      domain = CacheHelpers.insert_domain()
+
+      structure =
+        insert(:data_structure,
+          domain_ids: [domain.id],
+          rejected_note:
+            build(:structure_note,
+              df_content: df_content,
+              status: :rejected,
+              data_structure: nil
+            )
+        )
+
+      dsv =
+        insert(:data_structure_version,
+          data_structure: structure
+        )
+
+      %{id: id} =
+        CacheHelpers.insert_template(%{
+          name: dsv.type,
+          scope: "dd",
+          content: [
+            %{
+              "name" => "group",
+              "fields" => [
+                %{
+                  "name" => "string",
+                  "type" => "string",
+                  "label" => "Label foo"
+                }
+              ]
+            }
+          ]
+        })
+
+      insert(:data_structure_type, name: dsv.type, template_id: id)
+
+      assert :ok = StructureEnricher.refresh()
+
+      ElasticsearchMock
+      |> expect(:request, fn _, :post, "/structures/_search", _, opts ->
+        assert opts == [params: %{"scroll" => "1m"}]
+        SearchHelpers.scroll_response([dsv])
+      end)
+      |> expect(:request, fn _, :post, "/_search/scroll", body, [] ->
+        assert body == %{"scroll" => "1m", "scroll_id" => "some_scroll_id"}
+        SearchHelpers.scroll_response([])
+      end)
+
+      assert %{resp_body: body} =
+               post(
+                 conn,
+                 Routes.xlsx_path(conn, :download, %{
+                   download_type: "editable",
+                   note_type: "non_published"
+                 })
+               )
 
       assert {:ok, workbook} = XlsxReader.open(body, source: :binary)
       assert {:ok, [headers | content]} = XlsxReader.sheet(workbook, dsv.type)
@@ -127,7 +412,10 @@ defmodule TdDdWeb.DataStructures.XLSXControllerTest do
 
   describe "upload" do
     @tag authentication: [role: "admin"]
-    test "creates async job for upload processing", %{conn: conn, claims: claims} do
+    test "creates async job for upload processing for published notes", %{
+      conn: conn,
+      claims: claims
+    } do
       file = "test/fixtures/xlsx/upload.xlsx"
 
       opts = %{
@@ -144,7 +432,10 @@ defmodule TdDdWeb.DataStructures.XLSXControllerTest do
       hash = FileHash.hash("test/fixtures/xlsx/upload.xlsx", :md5)
 
       assert conn
-             |> post(Routes.xlsx_path(conn, :upload), structures: upload(file))
+             |> post(Routes.xlsx_path(conn, :upload),
+               structures: upload(file),
+               note_type: :published
+             )
              |> response(:accepted)
 
       assert_enqueued worker: UploadWorker,
@@ -162,6 +453,64 @@ defmodule TdDdWeb.DataStructures.XLSXControllerTest do
                filename: "upload.xlsx",
                task_reference: "oban:#{job_id}"
              }
+
+      assert_enqueued(
+        worker: UploadWorker,
+        args: %{hash: hash, opts: opts},
+        queue: :xlsx_upload_queue
+      )
+    end
+
+    @tag authentication: [role: "admin"]
+    test "creates async job for upload processing for non-published notes", %{
+      conn: conn,
+      claims: claims
+    } do
+      file = "test/fixtures/xlsx/upload.xlsx"
+
+      opts = %{
+        "auto_publish" => false,
+        "lang" => "en",
+        "user_id" => claims.user_id,
+        "claims" => %{
+          "user_id" => claims.user_id,
+          "user_name" => claims.user_name,
+          "jti" => claims.jti
+        }
+      }
+
+      hash = FileHash.hash("test/fixtures/xlsx/upload.xlsx", :md5)
+
+      assert conn
+             |> post(Routes.xlsx_path(conn, :upload),
+               structures: upload(file),
+               note_type: :non_published
+             )
+             |> response(:accepted)
+
+      assert_enqueued(
+        worker: UploadWorker,
+        args: %{hash: hash, opts: opts},
+        queue: :xlsx_upload_queue
+      )
+
+      assert [%Oban.Job{id: job_id}] = all_enqueued()
+
+      assert [event] = Repo.all(FileBulkUpdateEvent)
+
+      assert Map.take(event, [:user_id, :status, :hash, :filename, :task_reference]) == %{
+               user_id: claims.user_id,
+               status: "PENDING",
+               hash: hash,
+               filename: "upload.xlsx",
+               task_reference: "oban:#{job_id}"
+             }
+
+      assert_enqueued(
+        worker: UploadWorker,
+        args: %{hash: hash, opts: opts},
+        queue: :xlsx_upload_queue
+      )
     end
   end
 end